Refactor generators into separate modules
