+++ packages/labs/gen-wrapper-vue/.gitignore
@@ -1,7 +1,8 @@
 /test-gen/
 /test-output/
-/index.js
-/index.js.map
-/index.d.ts
-/index.d.ts.map
+/lib/
+/*.js
+/*.js.map
+/*.d.ts
+/*.d.ts.map
 /gen-output/
