+++ packages/labs/gen-wrapper-vue/src/lib/tsconfig-template.ts
@@ -1,0 +1,32 @@
+export const tsconfigTemplate = () => {
+  return JSON.stringify(
+    {
+      compilerOptions: {
+        target: 'es2019',
+        module: 'es2015',
+        lib: ['es2020', 'DOM', 'DOM.Iterable'],
+        declaration: true,
+        declarationMap: true,
+        sourceMap: true,
+        inlineSources: true,
+        outDir: './',
+        rootDir: './src',
+        strict: true,
+        noUnusedLocals: true,
+        noUnusedParameters: true,
+        noImplicitReturns: true,
+        noFallthroughCasesInSwitch: true,
+        noImplicitAny: true,
+        noImplicitThis: true,
+        moduleResolution: 'node',
+        allowSyntheticDefaultImports: true,
+        experimentalDecorators: true,
+        noImplicitOverride: true,
+      },
+      include: ['src/**/*.ts'],
+      exclude: [],
+    },
+    null,
+    2
+  );
+};
